+++ include/zephyr/net/socketcan.h
@@ -13,7 +13,6 @@
 #ifndef ZEPHYR_INCLUDE_NET_SOCKETCAN_H_
 #define ZEPHYR_INCLUDE_NET_SOCKETCAN_H_
 
-#include <zephyr/drivers/can.h>
 #include <zephyr/types.h>
 #include <zephyr/net/net_ip.h>
 #include <zephyr/net/net_if.h>
@@ -118,67 +117,6 @@ struct socketcan_filter {
 	/** The mask applied to @a can_id for matching. */
 	socketcan_id_t can_mask;
 };
-
-/**
- * @brief Translate a @a socketcan_frame struct to a @a can_frame struct.
- *
- * @param sframe Pointer to sockecan_frame struct.
- * @param zframe Pointer to can_frame struct.
- */
-static inline void socketcan_to_can_frame(const struct socketcan_frame *sframe,
-					  struct can_frame *zframe)
-{
-	zframe->id_type = (sframe->can_id & BIT(31)) >> 31;
-	zframe->rtr = (sframe->can_id & BIT(30)) >> 30;
-	zframe->id = sframe->can_id & BIT_MASK(29);
-	zframe->dlc = sframe->can_dlc;
-	memcpy(zframe->data, sframe->data, MIN(sizeof(sframe->data), sizeof(zframe->data)));
-}
-
-/**
- * @brief Translate a @a can_frame struct to a @a socketcan_frame struct.
- *
- * @param zframe Pointer to can_frame struct.
- * @param sframe  Pointer to socketcan_frame struct.
- */
-static inline void socketcan_from_can_frame(const struct can_frame *zframe,
-					    struct socketcan_frame *sframe)
-{
-	sframe->can_id = (zframe->id_type << 31) | (zframe->rtr << 30) | zframe->id;
-	sframe->can_dlc = zframe->dlc;
-	memcpy(sframe->data, zframe->data, MIN(sizeof(zframe->data), sizeof(sframe->data)));
-}
-
-/**
- * @brief Translate a @a socketcan_filter struct to a @a can_filter struct.
- *
- * @param sfilter Pointer to socketcan_filter struct.
- * @param zfilter Pointer to can_filter struct.
- */
-static inline void socketcan_to_can_filter(const struct socketcan_filter *sfilter,
-					   struct can_filter *zfilter)
-{
-	zfilter->id_type = (sfilter->can_id & BIT(31)) >> 31;
-	zfilter->rtr = (sfilter->can_id & BIT(30)) >> 30;
-	zfilter->id = sfilter->can_id & BIT_MASK(29);
-	zfilter->rtr_mask = (sfilter->can_mask & BIT(30)) >> 30;
-	zfilter->id_mask = sfilter->can_mask & BIT_MASK(29);
-}
-
-/**
- * @brief Translate a @a can_filter struct to a @a socketcan_filter struct.
- *
- * @param zfilter Pointer to can_filter struct.
- * @param sfilter Pointer to socketcan_filter struct.
- */
-static inline void socketcan_from_can_filter(const struct can_filter *zfilter,
-					     struct socketcan_filter *sfilter)
-{
-	sfilter->can_id = (zfilter->id_type << 31) |
-		(zfilter->rtr << 30) | zfilter->id;
-	sfilter->can_mask = (zfilter->rtr_mask << 30) |
-		(zfilter->id_type << 31) | zfilter->id_mask;
-}
 
 /** @} */
