+++ subsys/net/lib/sockets/sockets_can.c
@@ -21,6 +21,7 @@ LOG_MODULE_REGISTER(net_sock_can, CONFIG_NET_SOCKETS_LOG_LEVEL);
 #include <zephyr/sys/fdtable.h>
 #include <zephyr/net/canbus.h>
 #include <zephyr/net/socketcan.h>
+#include <zephyr/net/socketcan_utils.h>
 #include <zephyr/drivers/can.h>
 
 #include "sockets_internal.h"
